test working, no more alemdic

+++ backend/user_service/models.py
@@ -1,5 +1,5 @@
-from sqlalchemy import Column, Integer, String, Float
-from sqlalchemy.ext.declarative import declarative_base
+from sqlalchemy import Column, Integer, String, Float, Double
+from sqlalchemy.orm import declarative_base
 
 Base = declarative_base()
 
@@ -9,8 +9,8 @@ class User(Base):
     email = Column(String, unique=True, index=True)
     hashed_password = Column(String, nullable=True)
     age = Column(Integer, nullable=True)
-    height = Column(Float, nullable=True)
-    weight = Column(Float, nullable=True)
+    height = Column(Double, nullable=True)
+    weight = Column(Double, nullable=True)
     exercise_duration = Column(Integer, nullable=True)
     exercise_frequency = Column(Integer, nullable=True)
     exercise_goal = Column(Integer, nullable=True)
